CustomAdapterとGitItemを隔離
