+++ app/src/main/kotlin/jp/co/yumemi/android/codeCheck/CustomAdapter.kt
@@ -4,6 +4,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 
@@ -31,5 +32,15 @@ class CustomAdapter(
         holder.itemView.setOnClickListener {
             itemClickListener.itemClick(item)
         }
+    }
+}
+
+val diff_util = object : DiffUtil.ItemCallback<GitItem>() {
+    override fun areItemsTheSame(oldItem: GitItem, newItem: GitItem): Boolean {
+        return oldItem.name == newItem.name
+    }
+
+    override fun areContentsTheSame(oldItem: GitItem, newItem: GitItem): Boolean {
+        return oldItem == newItem
     }
 }
